+++ src/safeUI/Components/submitForm.tsx
@@ -26,8 +26,20 @@ import {
  * improve readability.
  */
 
+// --- Additional Feature - Jaafar ---
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+import { StyledButton, StyledSubmitButton } from '../Styles/Styled';
+import { TextField } from '@mui/material';
+const MAX_EMAIL_CHARACTERS = 256;
+// --- end ---
+
 const MAX_CHARACTERS = 7500;
 const MAX_Subject_CHARACTERS = 100;
+const port = 3001;
 
 function SubmitForm() {
   const [characterCount, setCharCount] = useState(0);
@@ -44,6 +56,99 @@ function SubmitForm() {
   const handleOpenError = () => setOpenError(true);
   const handleCloseError = () => setOpenError(false);
 
+  // --- Additional Feature - Jaafar ---
+  const [code, setCode] = useState('');
+  const [open, setOpen] = useState(false);
+  const [openEmail, setOpenEmail] = useState(false);
+  const [openCode, setOpenCode] = useState(false);
+  const [email, setEmail] = useState('');
+  const [emailError, setEmailError] = useState(false);
+  const [validEmail, setValidEmail] = useState(false);
+  const [emailHelperText, setEmailHelperText] = useState('');
+  const emailRegex = new RegExp(
+    /^[A-Za-z0-9_!#$%&'*+/=?`{|}~^.-]+@[A-Za-z0-9.-]+$/,
+    'gm'
+  );
+
+  const handleSnackbarOpen = () => {
+    setOpen(true);
+  };
+
+  const handleEmailChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    const newEmail = e.target.value;
+    const newCharacterCount = newEmail.length;
+    if (newCharacterCount <= MAX_EMAIL_CHARACTERS) {
+      setEmail(newEmail);
+      setValidEmail(emailRegex.test(newEmail));
+    }
+  };
+
+  const handleClose = () => {
+    setOpenEmail(false);
+    setOpenCode(false);
+    // refresh the page once user hit close button
+    window.location.reload();
+  };
+
+  const handleYes = () => {
+    fetch(`http://131.252.208.28:${port}/setReply`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        code: code,
+      }),
+    })
+      // response from fetch
+      .then((response) => {
+        if (!response.ok) {
+          throw new Error(response.statusText);
+        }
+        return response.text();
+      });
+    setOpenEmail(false);
+    setOpenCode(true);
+  };
+
+  const handleEmail = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    setEmailError(false);
+    setEmailHelperText('');
+    if (email === '') {
+      setEmailError(true);
+      setEmailHelperText('This field is required');
+    } else if (validEmail === false) {
+      setEmailError(true);
+      setEmailHelperText('Enter a valid email');
+    } else {
+      fetch(`http://131.252.208.28:${port}/receiverEmail`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          email: email,
+          code: code,
+        }),
+      })
+        // response from fetch
+        .then((response) => {
+          if (!response.ok) {
+            throw new Error(response.statusText);
+          }
+          window.location.reload();
+          return response.text();
+        })
+        .catch((error) => {
+          console.error('There was a problem with the fetch operation:', error);
+          handleOpenError();
+        });
+      setOpenCode(false);
+    }
+  };
+  // --- end ---
+
   useEffect(() => {
     if (shouldReload) {
       window.location.reload();
@@ -58,7 +163,6 @@ function SubmitForm() {
 
     if (to && subject && message) {
       // send POST request to 'addMessage' route
-      const port = 3001;
       fetch(`http://131.252.208.28:${port}/addMessage`, {
         method: 'POST',
         headers: {
@@ -76,14 +180,24 @@ function SubmitForm() {
           message_reply: null,
         }),
       })
+        // --- Additional feature - Jaafar ---
+        // response from fetch
         .then((response) => {
-          if (!response.ok) {
-            throw new Error('Network response was not ok');
-          } else {
-            setOpenSuccess(true);
+          if (response.status === 400) {
+            throw new Error('Invalid message: message contains profanities');
+          } else if (!response.ok) {
+            throw new Error(response.statusText);
           }
+          setOpenEmail(true);
+          handleSnackbarOpen();
+          return response.text();
         })
+        .then((responseText) => {
+          setCode(responseText);
+        })
+        // --- End ---
         .catch((error) => {
+          setOpenEmail(false);
           console.error('There was a problem with the fetch operation:', error);
           handleOpenError();
         });
@@ -130,6 +244,74 @@ function SubmitForm() {
           onClose={() => handleCloseSuccessSent(setSubject, setShouldReload)}
         />
       </Grid>
+
+      {/* --- Additonal Feature - Jaafar --- */}
+      <Dialog
+        open={openEmail}
+        onClose={handleClose}
+        aria-labelledby='responsive-dialog-title'
+      >
+        <DialogTitle id='responsive-dialog-title'>
+          Your feedback successfully sent to PSU CS Department! Would you like
+          to receive repliy to your feedback?
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            If yes, you will get a code that you can use to check back here for
+            reply
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <StyledButton autoFocus onClick={handleClose}>
+            No
+          </StyledButton>
+          <StyledButton onClick={handleYes} autoFocus>
+            Yes
+          </StyledButton>
+        </DialogActions>
+      </Dialog>
+
+      <Dialog
+        open={openCode}
+        onClose={handleClose}
+        aria-labelledby='responsive-dialog-title'
+      >
+        <DialogTitle id='responsive-dialog-title'>
+          {'Would you like to receive your code by Email?'}
+        </DialogTitle>
+        <form noValidate autoComplete='off' onSubmit={handleEmail}>
+          <DialogContent>
+            <DialogContentText>Code: {code}</DialogContentText>
+            <DialogContentText>
+              Here is your code. If you wish for us to email you the code please
+              enter your email. Your email will not be saved!
+            </DialogContentText>
+            <TextField
+              autoFocus
+              margin='dense'
+              id='name'
+              label='Email Address'
+              type='email'
+              fullWidth
+              variant='standard'
+              placeholder='Enter Email'
+              autoComplete='off'
+              spellCheck='false'
+              onChange={handleEmailChange}
+              error={emailError}
+              helperText={emailHelperText}
+              inputProps={{ maxlength: MAX_EMAIL_CHARACTERS }}
+            />
+          </DialogContent>
+          <DialogActions>
+            <StyledButton onClick={handleClose}>No</StyledButton>
+            <StyledSubmitButton variant='contained' type='submit'>
+              Submit
+            </StyledSubmitButton>
+          </DialogActions>
+        </form>
+      </Dialog>
+      {/* --- End --- */}
     </form>
   );
 }
